Improvements to oonidata CLI (#7)
Make use of the new measurement listing and download functions for the
CLI
This also makes significant changes to both the CLI arguments as well as
the output layout on disk.

Changes include:
* since and until are renamed to start-day and end-day. This is to avoid
confusion with the since and until used in OONI Explorer and API, which
are related to the the measurement_start_time, while here it's the
upload time
* country-codes is renamed to probe-cc and takes in a CSV list of
country codes. This is done for consistency with the API
* test-names is renamed to test-name for consistency with the API. It
also takes a CSV list of test names.
* The files are written to disk in a slightly different layout that
makes it easier to work with.

Improvements include:
* There are now progress bars that are a bit more accurate and useful

We also drop all the now unused code such as s3feeder, utils, mytypes,
metrics, etc.