Assess and improve performance of observation generation (#37)
Summary of changes:
* Parallelise on file entry instead of bucket_date. This leads to
significant performance boost for processing measurements in a day
* Record inside of `oonidata_processing_logs` metrics about how long it
took process a certain daily batch
* Change schema of HIRL to include sent and received
* Various dask related performance tuning
* Additional metric collection inside of statsd

Prior to this refactor we were achieving a rate of 5000 measurements per
second, but were only reaching this speed when processing multiple
bucket_dates at the same time.

This was happening because the parallelisation was happening on a per
day bucket basic. When processing a single day worth of data (which is
what happens in the daily batch), we were only reaching a speed of about
500 measurements per second.

As part of this PR I have added support for parallelising on a per
archived file basis, which allows us to reach a performance of 7-8k
measurements per second on a per day basis (which is actually faster
than what we were doing before). The splitting into batches happens
based on the size of the files so we should be getting a pretty
consistent performance irrespective of the size of the daily batch.

This is the first step towards improving the performance of reprocessing
of legacy buckets, where each daily bucket is too small to benefit from
just splitting based on day (we ideally want to process batches of days
together).

I also implemented as part of this same PR some changes to the schema of
HIRL to make it more useful for analysis.