feat: introduce address based method for events
This diff introduces a new address based method for
finding network events' windows for specific tls
handshakes. This is done by using the address field in
the `tls_handshake` as a key and filtering the `network_events`
using this key. On the off chance that we get two
windows for same addresses, we return the first one. This
is done to tackle the presence of `transaction_id` in `network_events`
which prevents us from assuming that the events are in order
and can be iterated as is.