Fix bug in parsing TLS handshake time events
Some tests, such as whatsapp, signal and telegram, do not begin with the
connect network event, but also don't make use of the transaction_id.

This resulted in parsing issues where the start and end time of network
events were inconsistent.

We change the logic to only parse network events relevant to TLS
handshakes within the connect and tls_handshake_done boundaries.